Django backend/APIProject/requirements/requirements.txt
Django==3.1.7
coverage==5.3.0
djangorestframework==3.12.2
markdown==3.3.4
django-filter==2.4.0